==> main/2.5-Data-Cleaning-PartTwo.ipynb <==
# More Data Cleaning 

We realised that the document-term matrices we created in 2-Data-Cleaning.ipynb using Count Vectorizer and TF-IDF Vectorizer has a lot of meaningless filler words and common words such as `'like'`, `'just'`, `'people'`, `'youre'` and etc. 

Therefore, we wish to inspect the matrices further and create a new stop words list in this notebook. 

import pandas as pd
from collections import Counter
import pickle
import re
import string 
import nltk
import scipy
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords 
from nltk.stem import WordNetLemmatizer 
from nltk.stem import PorterStemmer
from nltk.corpus import wordnet
from sklearn.feature_extraction import text 
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer

# read in the document-term matrix formed by Count Vectorizer 
df_cv = pd.read_pickle('/Users/lihuicham/Documents/GitHub/standup-comedy-analysis/pickle/cv.pkl')
# transpose to term-document matrix 
df_cv = df_cv.transpose()   

# Find the top 50 words in each transcript 
top_dict = {}
for c in df_cv.columns:
    top = df_cv[c].sort_values(ascending=False).head(50)
    top_dict[c]= list(zip(top.index, top.values))

# Print the top 50 words in each transcript 
# for transcript, top_words in top_dict.items():
#     print(transcript)
#     print(', '.join([word for word, count in top_words[0:49]]))
#     print('---')

# we add the most common top words to a stop word list

# Let's first pull out the top 50 words for each comedian
words = []
for transcript in df_cv.columns:
    top = [word for (word, count) in top_dict[transcript]]
    for t in top:
        words.append(t)

# Aggregate this list and identify the most common words along with how many transcripts they occur in
most_common_words = Counter(words).most_common()

# create our own stop word list based on top words 
# we consider the word as a stop word if >= 150 transcript have it as top word

add_stop_words = [word for word, count in most_common_words if count >= 150]

# pickle
with open('/Users/lihuicham/Documents/GitHub/standup-comedy-analysis/pickle/' + 'mostcommonwords-st.pkl', 'wb') as f:
    pickle.dump(most_common_words, f)

# after a few iterations of checking the top words with Count Vectorizer
# we created a list of stop words that needs to be removed too

own_stop_words = ['just', 'okay', 'ive', '♪', '–', 'ta', 'uh', 'wan', 'g', 'e', 'ah', 'r', 'mi', 'le']
complete_stop_words = [*add_stop_words, *own_stop_words]
complete_stop_words

## Helper Functions 
From 2-Data-Cleaning.ipynb file. 

# same function as 2-Data-Cleaning 
def get_wordnet_pos(treebank_tag) : 
    if treebank_tag.startswith('J'):
        return wordnet.ADJ
    elif treebank_tag.startswith('V'):
        return wordnet.VERB
    elif treebank_tag.startswith('N'):
        return wordnet.NOUN
    elif treebank_tag.startswith('R'):
        return wordnet.ADV
    else:
        # As default pos in lemmatization is Noun
        return wordnet.NOUN

# same function as 2-Data-Cleaning 

lemmatizer = WordNetLemmatizer()

def pos_then_lemmatize(pos_tagged_words) :
    res = []
    for pos in pos_tagged_words : 
        word = pos[0]
        pos_tag = pos[1]

        lem = lemmatizer.lemmatize(word, get_wordnet_pos(pos_tag))
        res.append(lem)
    return res

# new function 
def custom_tokenizer_stop(doc) : 
    words = word_tokenize(doc.lower())
    
    # add our own stop word list to the existing English stop words 
    new_stop_words = text.ENGLISH_STOP_WORDS.union(complete_stop_words)
    
    filtered_words = [w for w in words if not w in new_stop_words] 
    pos_tagged_words = nltk.pos_tag(filtered_words)
    pos_lemmatized_words = pos_then_lemmatize(pos_tagged_words)
    filtered_words_2 = [w for w in pos_lemmatized_words if not w in new_stop_words] 
    
    return filtered_words_2

## An updated Document-Term Matrix 

### Count Vectorizer

# read in the clean data 
df_clean = pd.read_pickle('/Users/lihuicham/Documents/GitHub/standup-comedy-analysis/pickle/corpus.pkl')
df_clean.head()

from sklearn.feature_extraction.text import CountVectorizer

# (1, 2) : include bigram 
# max_features = 300 : choose features/words that occur most frequently to be its vocabulary 
cv = CountVectorizer(ngram_range = (1, 1),
                    tokenizer = custom_tokenizer_stop)
cv_vectors = cv.fit_transform(df_clean['Transcript'])
cv_feature_names = cv.get_feature_names_out()
cv_matrix_stop = pd.DataFrame(cv_vectors.toarray(), columns=cv_feature_names)

### Double Checking

In the below code chunk, we double check whether our `completed_stop_words` list is working.  

In `top_dict_check`, we can clearly see that now the words are starting to makes sense and are indeed meaningful in each transcript. The common top words that are meaningless and filler words are removed successfully. 

# we double check on the top words in each transcript now. 
cv_matrix_check = cv_matrix_stop.transpose()

top_dict_check_cv = {}
for c in cv_matrix_check.columns:
    top = cv_matrix_check[c].sort_values(ascending=False).head(30)
    top_dict_check_cv[c]= list(zip(top.index, top.values))

# pickle 
with open('/Users/lihuicham/Documents/GitHub/standup-comedy-analysis/pickle/' + 'common_words_cv.pkl', 'wb') as f:
    pickle.dump(top_dict_check_cv, f)

# we check with the first transcript 
first_transcript_value_cv = list(top_dict_check_cv.values())[0]
first_transcript_value_cv 

# pickle the updated document-term matrix from Count Vectorizer
with open('/Users/lihuicham/Documents/GitHub/standup-comedy-analysis/pickle/' + 'cv_stop.pkl', 'wb') as f:
    pickle.dump(cv_matrix_stop, f)

### TF-IDF

We do the same for TF-IDF too.  
Output : An updated TF-IDF matrix

# TF-IDF Vectorizer

tf = TfidfVectorizer(ngram_range = (1, 1),
                    tokenizer = custom_tokenizer_stop)
tf_vectors = tf.fit_transform(df_clean['Transcript'])
tf_feature_names = tf.get_feature_names_out()
tfidf_matrix_stop = pd.DataFrame(tf_vectors.toarray(), columns=tf_feature_names)

# we double check on the top words in each transcript now. 
tf_matrix_check = tfidf_matrix_stop.transpose()

top_dict_check_tf = {}
for c in tf_matrix_check.columns:
    top = tf_matrix_check[c].sort_values(ascending=False).head(30)
    top_dict_check_tf[c]= list(zip(top.index, top.values))

# pickle 
with open('/Users/lihuicham/Documents/GitHub/standup-comedy-analysis/pickle/' + 'common_words_tfidf.pkl', 'wb') as f:
    pickle.dump(top_dict_check_tf, f)

first_transcript_value_tf = list(top_dict_check_tf.values())[0]
first_transcript_value_tf 

with open('/Users/lihuicham/Documents/GitHub/standup-comedy-analysis/pickle/' + 'tfidf_stop.pkl', 'wb') as f:
    pickle.dump(tfidf_matrix_stop, f)

## Decision Making 

Now, we need to decide which document-term matrix to use for the project.  
1. Count Vectorizer 
2. TF-IDF Vectorizer 

From the top words shown, **TF-IDF** might be a better matrix.  

Reasons : 
* More meaningful words that are useful for topic modelling and EDA. For example, important nouns such as `'kardashian'`, `'trimester'` and `'victim'` are valued in TF-IDF matrix compared to Count Vectorizer matrix. These words are important for topic modelling. 


# creating a sparse matrix for tfidf_stop 
sparse_tf_stop = scipy.sparse.csr_matrix(tfidf_matrix_stop.values)
tf_colnames = list(tfidf_matrix_stop.columns.values)

with open('/Users/lihuicham/Documents/GitHub/standup-comedy-analysis/pickle/' + 'sparse_tf_stop.pkl', 'wb') as f:
    pickle.dump(sparse_tf_stop, f)
    
with open('/Users/lihuicham/Documents/GitHub/standup-comedy-analysis/pickle/' + 'tf_colnames.pkl', 'wb') as f:
    pickle.dump(tf_colnames, f)
