==> main/4.5-Sentiment-Analysis-Classification.ipynb <==
import nltk
import pandas as pd 
from nltk.corpus import stopwords

# load pickle for most common words 
most_common_words = pd.read_pickle('/Users/lihuicham/Desktop/Y2S2/BT4222/project/standup-comedy-analysis/main/pickle/common_words_tfidf.pkl')

# load pickle for textblob dataframe 
df_textblob = pd.read_pickle('/Users/lihuicham/Desktop/Y2S2/BT4222/project/standup-comedy-analysis/main/pickle/sentiments_textblob.pkl')
print(df_textblob['Polarity_Class'].value_counts())

# load pickle for vader dataframe 
df_vader = pd.read_pickle('/Users/lihuicham/Desktop/Y2S2/BT4222/project/standup-comedy-analysis/main/pickle/sentiments_vader.pkl')
print(df_vader['Polarity_Class'].value_counts())

## TextBlob

df_textblob.head()

## Feature Extraction 

# get indexes of the positive and negative classes
pos_indexes = df_textblob.index[df_textblob['Polarity_Class'] == 1].tolist()
neg_indexes = df_textblob.index[df_textblob['Polarity_Class'] == 0].tolist()

# find out how many positive and negative words we have in total, along with their frequency (word, frequency)
pos_words_tuple = []
neg_words_tuple = []

for key in pos_indexes : 
    word_list = most_common_words[key]
    for word in word_list :
        pos_words_tuple.append(word)
        
for key in neg_indexes : 
    word_list = most_common_words[key]
    for word in word_list :
        neg_words_tuple.append(word)
        
print(f"Number of Positive Words : {len(pos_words_tuple)}" )
print(f"Number of Negative Words : {len(neg_words_tuple)}" )

# sort the based on descending frequency  (most frequent to least frequent)
sorted_pos_words = sorted(pos_words_tuple, key=lambda w: w[1], reverse=True)
sorted_neg_words = sorted(neg_words_tuple, key=lambda w: w[1], reverse=True)

pos_words = [w[0] for w in sorted_pos_words]
neg_words = [w[0] for w in sorted_neg_words]

# create UNIQUE positive and negative word list 

# first, find the common words between both lists
common_list = list(set(pos_words) & set(neg_words))
print(f"There are {len(common_list)} words that exists in both positive and negative word list.") 

# remove the common words to create unique lists 
pos_unique_words = [w for w in pos_words if w not in common_list]
neg_unique_words = [w for w in neg_words if w not in common_list]

print(f"Number of Unique Positive Words : {len(pos_unique_words)}" )
print(f"Number of Unique Negative Words : {len(neg_unique_words)}" )

# more cleaning : remove stop words
# stop_words = set(stopwords.words('english'))
# filtered_pos_unique_words = [w for w in pos_unique_words if w not in stop_words]
# filtered_neg_unique_words = [w for w in neg_unique_words if w not in common_list]

# print(f"Number of Filtered Unique Positive Words : {len(filtered_pos_unique_words)}" )   # 6579
# print(f"Number of Filtered Unique Negative Words : {len(filtered_neg_unique_words)}" )   # 573

# from this result, we know that removing stop words is not very useful 

# notice that there are actually a lot of repeated word in each list, 
# especially in the positive words list. 
# let's try to remove the duplicates 
# every word should only exist once within its own list (no duplicates)

# Note : we cannot use the set() function to remove duplicate as the method does not preserve the sorted order. 

pos_no_dup = []
neg_no_dup = []

for word in pos_unique_words:
    if word not in pos_no_dup:
        pos_no_dup.append(word)

for word in neg_unique_words:
    if word not in neg_no_dup:
        neg_no_dup.append(word)

print(f"Number of Unique Positive Words with No Duplicates : {len(pos_no_dup)}" )
print(f"Number of Unique Negative Words with No Duplicates : {len(neg_no_dup)}" )

own_stop_words = ['♫', 'fucken', 'motherfcker', 'whaaaaaaaa', 
                  'whaaaaaaaaaaaaaaaa', 'mmmmm', 'i–i', 'up♪', 'like—', '♪girl',
                 'yeah♪', 'you-', ]

filtered_pos_words = [w for w in pos_no_dup if w not in own_stop_words]
filtered_neg_words = [w for w in neg_no_dup if w not in own_stop_words]

print(f"Number of Filtered Positive Words : {len(filtered_pos_words)}" )  
print(f"Number of Filtered Negative Words : {len(filtered_neg_words)}" )
