==> main/6-Topic-Modeling.ipynb <==
# Topic Modeling with Empath 

## Empath module information 
GitHub Repo : [Empath](https://github.com/Ejhfast/empath-client)  
Research publication : [Empath : Understanding Topic Signals in Large-Scale Text](https://arxiv.org/pdf/1602.06979.pdf)



We will compare three methods used for Topic Modeling 
1. **Document-Term Matrix** from TF-IDF Vectorizer  
    Text is preprocessed (no stop words, lemmatization, parts-of-speech tagging and etc.) 
    Bag-of-Words where order of words is not preserved   

2. **Corpus** (aka. full transcript)  
   Order of words is preserved  

3. **Subtitle**  
   The original subtitle of each transcript, typically consists of one or two lines of summary of the transcript. 

import pandas as pd
from empath import Empath

lexicon = Empath()

### 1. Topic Modeling with Document-Term Matrix 

# load the top 30 common words dictionary created using TF-IDF vectorizer
common_words_dict = pd.read_pickle('/Users/lihuicham/Desktop/Y2S2/BT4222/project/standup-comedy-analysis/main/pickle/common_words_tfidf.pkl')

words = []
for i in common_words_dict[32] : 
    words.append(i[0])
    
sent = ' '.join(words)
sent

lext_dict_dtm = lexicon.analyze(sent, normalize=True)

# top 15 categories 
dtm_topics = sorted(lext_dict_dtm, key=lext_dict_dtm.get, reverse=True)[:15]

### 2. Topic Modeling with Corpus (Full Transcript) 

# load corpus 
df = pd.read_pickle('/Users/lihuicham/Desktop/Y2S2/BT4222/project/standup-comedy-analysis/main/pickle/corpus.pkl')
df.head()

transcript = df.Transcript[32].strip()
# transcript

lext_dict_transcript = lexicon.analyze(transcript, normalize=True)

# top 20 categories 
corpus_topics = sorted(lext_dict_transcript, key=lext_dict_transcript.get, reverse=True)[:20]

### 3. Topic Modeling with Subtitle 

subtitle = df.Subtitle[32]
subtitle

lext_dict_subtitle = lexicon.analyze(subtitle, normalize=True)

# top 10 categories 
subtitle_topics = sorted(lext_dict_subtitle, key=lext_dict_subtitle.get, reverse=True)[:10]

## Comparing the three methods 

corpus_topics  # 20 topics 

dtm_topics  # 15 topics 

subtitle_topics  # 10 topics 

### Getting the common topics by pairing the methods 

# corpus X dtm 
set_corpus_dtm = set(corpus_topics)&set(dtm_topics) 
corpus_dtm = sorted(set_corpus_dtm, key = lambda k : corpus_topics.index(k))
corpus_dtm

# dtm X subtitle 
set_dtm_subtitle = set(dtm_topics)&set(subtitle_topics) 
dtm_subtitle = sorted(set_dtm_subtitle, key = lambda k : dtm_topics.index(k))
dtm_subtitle

# corpus X subtitle  
set_corpus_subtitle = set(corpus_topics)&set(subtitle_topics) 
corpus_subtitle = sorted(set_corpus_subtitle, key = lambda k : corpus_topics.index(k))
corpus_subtitle

# corpus X dtm X subtitle 
set_all_methods = set(corpus_topics)&set(subtitle_topics)&set(dtm_topics) 
all_methods = sorted(set_corpus_subtitle, key = lambda k : corpus_topics.index(k))
all_methods

## Conclusion 

We can consider combining all three methods and produce one or two most possible topic(s) mentioned in each transcript. 

## Test 
Use one simple sentence to test with the `Empath` library

lex_dict_test = lexicon.analyze("he hit the other person", normalize=True)
# uncomment below line to see the normalized counts of each categories 
# lex_dict_test 

# get the top 5 categories 
sorted(lex_dict_test, key=lex_dict_test.get, reverse=True)[:5]
