==> main/7-Pipeline-Preparation.ipynb <==
import pandas as pd
from textblob import TextBlob
from pipeline_funcs import *
import pickle

df = pd.read_csv('transcripts.csv')
df

valid_transcripts_sent_df = pd.read_pickle('chloe_valid_transcripts_sent_df')
valid_transcripts_sent_df

# get all the valid titles from 'valid_transcripts_sent_df'
valid_titles = valid_transcripts_sent_df['Title'].unique()
valid_titles

#filter df for the rows with title in 'valid_titles'
valid_transcripts_df = df[df['Title'].isin(valid_titles)]
valid_transcripts_df

#find the 2 extra rows in 'valid_transcripts_df' and remove them
remove_idx1 = valid_transcripts_df[(valid_transcripts_df['Comedian'] == 'Dave Chappelle') & (valid_transcripts_df['Title'] == 'SNL Monologue (2020) – Transcript')].index[0]
remove_idx2 = valid_transcripts_df[(valid_transcripts_df['Comedian'] == 'Chris Rock') & (valid_transcripts_df['Title'] == 'SNL Monologue (2020) – Transcript')].index[0]

valid_transcripts_df = valid_transcripts_df.drop([remove_idx1, remove_idx2])
valid_transcripts_df

#add column for cleaned transcripts
valid_transcripts_df['Cleaned Transcript'] = valid_transcripts_df['Transcript'].apply(preprocess)
valid_transcripts_df

#find polarity and subjectivity for each cleaned transcript in 'valid_transcripts_df'
polarity_col = []
subjectivity_col = []

def get_pol_sub(text):
    blob = TextBlob(text)
    polarity_col.append(blob.sentiment.polarity)
    subjectivity_col.append(blob.sentiment.subjectivity)

valid_transcripts_df['Cleaned Transcript'].apply(get_pol_sub)

#add polarity and subjectivity column to 'valid_transcripts_df'
valid_transcripts_df['Polarity'] = polarity_col
valid_transcripts_df['Subjectivity'] = subjectivity_col
valid_transcripts_df

pickle.dump(valid_transcripts_df, open('valid_transcripts_df', 'wb'))
