==> main/7.5-Pipeline.ipynb <==
#user_input = input('Input the comedy transcript you would like to predict: ') #use this to take in real-time input and run the pipeline
user_input = open('test_pipeline_transcript.txt', 'r').read()

user_input = ' '.join(user_input.split('\n'))

import numpy as np
import pandas as pd
from pipeline_funcs import *
from imblearn.over_sampling import RandomOverSampler
import pickle

valid_transcripts_sent_df = pd.read_pickle('chloe_valid_transcripts_sent_df')
valid_transcripts_sent_df

valid_transcripts_df = pd.read_pickle('valid_transcripts_df')
valid_transcripts_df

#filter transcripts in 'valid_transcripts_df' for those similar to 'user_input' in terms of polarity and subjectivity
filtered_transcripts_df = filter_by_sentiment(valid_transcripts_df, user_input)
filtered_transcripts_df

#find the rows in 'valid_transcripts_sent_df' whose title is in 'filtered_transcripts_df' 
filtered_transcripts_sent_df = valid_transcripts_sent_df[valid_transcripts_sent_df['Title'].isin(filtered_transcripts_df['Title'])]
filtered_transcripts_sent_df

#get model that is pre-trained on all valid transcripts, and also the vectorizer, and features
full_grad_clf = pickle.load(open('trained_grad_model', 'rb'))
full_vectorizer = pickle.load(open('full_vectorizer', 'rb'))
full_features = pickle.load(open('full_features', 'rb'))

#TF-IDF vectorization on 'filtered_transcripts_sent_df' using full_vectorizer
filtered_tfidf_matrix = full_vectorizer.transform(filtered_transcripts_sent_df['Processed Transcript'])

X_train_filtered = pd.DataFrame(filtered_tfidf_matrix.toarray(), columns = full_vectorizer.get_feature_names_out())
X_train_filtered = X_train_filtered[full_features]
y_train_filtered = filtered_transcripts_sent_df['Funniness']

#get counts of each funniness in 'y_train_filtered'
y_train_filtered.value_counts()

#oversample to account for inbalanced data
oversampler = RandomOverSampler(sampling_strategy = 'minority')
X_train_sm, y_train_sm = oversampler.fit_resample(X_train_filtered, y_train_filtered)
y_train_sm.value_counts()

#train gradient boosting classifier model on X_train_sm, y_train_sm
from sklearn.ensemble import GradientBoostingClassifier

filtered_grad_clf = GradientBoostingClassifier(n_estimators = 40, learning_rate = 0.1, random_state = 0).fit(X_train_sm, y_train_sm)

#function to get predictions weighted by 'full_grad_clf' and 'filtered_grad_clf'
def weighted_predictions(full_model, filtered_model, X, weight = 0.75):
    full_probs = full_model.predict_proba(X)
    filtered_probs = filtered_model.predict_proba(X)

    #determine if filtered_model has fitted to all classes
    unfitted_classes = []
    for i in range(4):
        if i not in filtered_model.classes_:
            unfitted_classes.append(i)

    #add column of zero to filtered_probs for each unfitted class
    if len(unfitted_classes) > 0:
        for c in unfitted_classes:
            filtered_probs = np.insert(filtered_probs, c, 0, axis = 1)

    adj_probs = (full_probs * weight) + (filtered_probs * (1 - weight))
    y_hat = adj_probs.argmax(axis = 1)
    return y_hat

#break 'user_input' down into 10 chunks
def break_chunks(input, n_chunks):
    user_tokens = [s for s in input]
    user_token_count = len(user_tokens)
    chunk_token_count = user_token_count // n_chunks
    chunk_ranges = range(0, user_token_count, chunk_token_count)
    chunks = []

    for c in range(n_chunks):
        if c == n_chunks - 1:
            chunk = ''.join(user_tokens[chunk_ranges[c]:])
        else:
            chunk = ''.join(user_tokens[chunk_ranges[c]:chunk_ranges[c + 1]])
        chunks.append(chunk)

    return chunks

user_chunks = break_chunks(user_input, 10) 

#preprocess each chunk in 'user_chunks' and use the 'full_vectorizer' to transform it
chunk_df = pd.DataFrame({'Chunk': user_chunks})
chunk_df['Processed Chunk'] = chunk_df['Chunk'].apply(preprocess)
vec_user_input = full_vectorizer.transform(chunk_df['Processed Chunk'])
user_input_df = pd.DataFrame(vec_user_input.toarray(), columns = full_vectorizer.get_feature_names_out())
user_input_df = user_input_df[full_features]

#make predictions of funniness of 'user_input' 
'''
warm_pred_prob = full_grad_clf.predict_proba(user_input_df)
warm_pred = np.argmax(warm_pred_prob, axis = 1)
'''
weighted_pred = weighted_predictions(full_grad_clf, filtered_grad_clf, user_input_df)

#add column for predictions of funniness
chunk_df['Funniness'] = weighted_pred
chunk_df

#function to output funniness prediction in word form
def get_pred_output(pred):
    if pred == 0:
        return 'Neutral'
    elif pred == 1:
        return 'A little funny'
    elif pred == 2:
        return 'Moderately funny'
    elif pred == 3:
        return 'Very funny'

chunk_df['Funniness Word'] = chunk_df['Funniness'].apply(get_pred_output)
chunk_df

#topic modelling of similar transcripts
from empath import Empath
lexicon = Empath()

def get_topics(text, n_topics):
    topics_lst = []
    top_dict = lexicon.analyze(text, normalize = True)
    sorted_top_lst = sorted(top_dict.items(), key = lambda x: x[1], reverse = True)

    for top_weight in sorted_top_lst[:n_topics]:
        topics_lst.append(top_weight[0])
    return topics_lst

filtered_transcripts_df['Topics'] = filtered_transcripts_df['Cleaned Transcript'].apply(get_topics, args = (3,))
filtered_transcripts_df

#topic modelling of user input chunks
chunk_df['Topics'] = chunk_df['Processed Chunk'].apply(get_topics, args = (3,))
chunk_df

#topic modelling of entire user input
user_topics = get_topics(preprocess(user_input), 3)
user_topics
